refactor(react): reworked cardgroup component
