--- packages/react/src/components/Cards/CardGroup/index.tsx
@@ -1,3 +1,0 @@
-export { default as CardGroup } from "./CardGroup";
-export type { CardGroupProps } from "./CardGroup.props";
-export { default as CardGroupArgs } from "./CardGroup.args";
