--- packages/react/src/components/Cards/DataCard/index.tsx
@@ -1,1 +1,0 @@
-export * from "./DataCard";
