+++ packages/react/src/components/Cards/DetailCard.tsx
@@ -1,13 +1,9 @@
-import { forwardRef } from "react";
 import classNames from "classnames";
+import { forwardRef } from "react";
+import useGlobalSettings from "../../hooks/useGlobalSettings";
+import { CardSize, EventDate, HeadingTypes, ThemeTypes } from "../../types";
 
-import useGlobalSettings from "../../../hooks/useGlobalSettings";
-import { LinkList, LinkListProps } from "../../LinkList";
-import { CardAlignment, CardSize, HeadingTypes } from "../../../types";
-
-export type MultiLinkCardProps = {
-  isVideo?: boolean;
-
+export type DetailCardProps = {
   /**
    * A line of text that appears as a small heading above the title of the card
    */
@@ -24,55 +20,66 @@ export type MultiLinkCardProps = {
   titleLevel?: HeadingTypes;
 
   /**
+   * Will render the card to appear on light or dark backgrounds
+   */
+  theme?: ThemeTypes;
+
+  /**
    * How big should the card be
    */
-  size?: CardSize;
+  size?: Omit<CardSize, "standard">;
 
   /**
-   * How should the card be aligned
+   * Specify the event Date, in both human and Unix format.
    */
-  alignment?: CardAlignment;
+  date?: EventDate;
 
   /**
-   * Introductory text in the card
+   * Additional date information
+   */
+  details?: string;
+
+  /**
+   * Intro text for the card
    */
   intro?: string;
 
   /**
-   * Link to the card
+   * Specify the URL for the card link
    */
   link?: string;
 
   /**
-   * Link list to show in the card
-   */
-  linklist?: LinkListProps;
-
-  /**
-   * The image to show in the card
+   * Specify the image for the card
    */
   image?: string;
 
   /**
-   * Specify an optional className to be added to your TextCard.
+   * Specify if the card is a video
+   */
+  isVideo?: boolean;
+
+  /**
+   * Specify an optional className to be added to your DetailCard.
    */
   className?: string;
 };
 
-const MultiLinkCard = forwardRef<HTMLDivElement, MultiLinkCardProps>(
+const DetailCard = forwardRef<HTMLDivElement, DetailCardProps>(
   (
     {
       className,
       title,
-      size = "standard",
-      alignment,
-      intro,
+      theme = "light",
+      size = "narrow",
+      date,
+      details,
       link,
-      linklist,
-      image,
       titleLevel: TitleElement = "p",
       eyebrow,
-      isVideo,
+      intro,
+      image,
+      isVideo = false,
     },
     ref
   ) => {
@@ -82,14 +89,13 @@ const MultiLinkCard = forwardRef<HTMLDivElement, MultiLinkCardProps>(
 
     const cardClasses = classNames(
       baseClass,
+      `${baseClass}__type__detail`,
       className,
-      `${baseClass}__type__multilink`,
       {
-        [`${baseClass}__align__${alignment}`]: alignment,
         [`${baseClass}__action`]: link,
-        [`${baseClass}__size__${size}`]: size,
+        [`${baseClass}__size__${String(size)}`]: size,
+        [`${baseClass}__theme__${theme}`]: theme,
         [`${baseClass}__isvideo`]: isVideo,
-        [`${baseClass}--no-image`]: !image,
       }
     );
 
@@ -105,36 +111,27 @@ const MultiLinkCard = forwardRef<HTMLDivElement, MultiLinkCardProps>(
             <div className={`${baseClass}--image--wrapper`}>
               <picture>
                 <img
+                  className={`${baseClass}--picture`}
                   src={image}
                   alt={title}
-                  className={`${baseClass}--picture`}
                 />
               </picture>
             </div>
           )}
           <div className={`${baseClass}--content`}>
             {eyebrow && <p className={`${baseClass}--eyebrow`}>{eyebrow}</p>}
-            <TitleElement className={`${baseClass}--title`}>
-              {title}
-            </TitleElement>
-            {image && (
-              <div className={`${baseClass}--image--wrapper`}>
-                <picture>
-                  <img
-                    src={image}
-                    alt={title}
-                    className={`${baseClass}--picture`}
-                  />
-                </picture>
-              </div>
+            {title && (
+              <TitleElement className={`${baseClass}--title`}>
+                {title}
+              </TitleElement>
             )}
             {intro && <p className={`${baseClass}--intro`}>{intro}</p>}
-            {linklist && (
-              <LinkList
-                headline={linklist.headline}
-                linkgroup={linklist.linkgroup}
-              />
+            {date && (
+              <time className={`${baseClass}--date`} dateTime={date.unix}>
+                {date.human}
+              </time>
             )}
+            {details && <p className={`${baseClass}--date-extra`}>{details}</p>}
           </div>
         </div>
       </div>
@@ -142,4 +139,4 @@ const MultiLinkCard = forwardRef<HTMLDivElement, MultiLinkCardProps>(
   }
 );
 
-export { MultiLinkCard };
+export { DetailCard };
