--- packages/react/src/components/Cards/DetailCard/index.tsx
@@ -1,1 +1,0 @@
-export * from "./DetailCard";
