--- packages/react/src/components/Cards/FactListCard/index.tsx
@@ -1,1 +1,0 @@
-export * from "./FactListCard";
