--- packages/react/src/components/Cards/FeatureCard/index.tsx
@@ -1,1 +1,0 @@
-export * from "./FeatureCard";
