+++ packages/react/src/components/Cards/MultiLinkCard.tsx
@@ -1,9 +1,13 @@
-import classNames from "classnames";
 import { forwardRef } from "react";
-import useGlobalSettings from "../../../hooks/useGlobalSettings";
-import { CardSize, EventDate, HeadingTypes, ThemeTypes } from "../../../types";
+import classNames from "classnames";
 
-export type DetailCardProps = {
+import useGlobalSettings from "../../hooks/useGlobalSettings";
+import { LinkList, LinkListProps } from "../LinkList";
+import { CardAlignment, CardSize, HeadingTypes } from "../../types";
+
+export type MultiLinkCardProps = {
+  isVideo?: boolean;
+
   /**
    * A line of text that appears as a small heading above the title of the card
    */
@@ -20,66 +24,55 @@ export type DetailCardProps = {
   titleLevel?: HeadingTypes;
 
   /**
-   * Will render the card to appear on light or dark backgrounds
-   */
-  theme?: ThemeTypes;
-
-  /**
    * How big should the card be
    */
-  size?: Omit<CardSize, "standard">;
+  size?: CardSize;
 
   /**
-   * Specify the event Date, in both human and Unix format.
+   * How should the card be aligned
    */
-  date?: EventDate;
+  alignment?: CardAlignment;
 
   /**
-   * Additional date information
-   */
-  details?: string;
-
-  /**
-   * Intro text for the card
+   * Introductory text in the card
    */
   intro?: string;
 
   /**
-   * Specify the URL for the card link
+   * Link to the card
    */
   link?: string;
 
   /**
-   * Specify the image for the card
+   * Link list to show in the card
+   */
+  linklist?: LinkListProps;
+
+  /**
+   * The image to show in the card
    */
   image?: string;
 
   /**
-   * Specify if the card is a video
-   */
-  isVideo?: boolean;
-
-  /**
-   * Specify an optional className to be added to your DetailCard.
+   * Specify an optional className to be added to your TextCard.
    */
   className?: string;
 };
 
-const DetailCard = forwardRef<HTMLDivElement, DetailCardProps>(
+const MultiLinkCard = forwardRef<HTMLDivElement, MultiLinkCardProps>(
   (
     {
       className,
       title,
-      theme = "light",
-      size = "narrow",
-      date,
-      details,
+      size = "standard",
+      alignment,
+      intro,
       link,
+      linklist,
+      image,
       titleLevel: TitleElement = "p",
       eyebrow,
-      intro,
-      image,
-      isVideo = false,
+      isVideo,
     },
     ref
   ) => {
@@ -89,13 +82,14 @@ const DetailCard = forwardRef<HTMLDivElement, DetailCardProps>(
 
     const cardClasses = classNames(
       baseClass,
-      `${baseClass}__type__detail`,
       className,
+      `${baseClass}__type__multilink`,
       {
+        [`${baseClass}__align__${alignment}`]: alignment,
         [`${baseClass}__action`]: link,
-        [`${baseClass}__size__${String(size)}`]: size,
-        [`${baseClass}__theme__${theme}`]: theme,
+        [`${baseClass}__size__${size}`]: size,
         [`${baseClass}__isvideo`]: isVideo,
+        [`${baseClass}--no-image`]: !image,
       }
     );
 
@@ -111,27 +105,36 @@ const DetailCard = forwardRef<HTMLDivElement, DetailCardProps>(
             <div className={`${baseClass}--image--wrapper`}>
               <picture>
                 <img
-                  className={`${baseClass}--picture`}
                   src={image}
                   alt={title}
+                  className={`${baseClass}--picture`}
                 />
               </picture>
             </div>
           )}
           <div className={`${baseClass}--content`}>
             {eyebrow && <p className={`${baseClass}--eyebrow`}>{eyebrow}</p>}
-            {title && (
-              <TitleElement className={`${baseClass}--title`}>
-                {title}
-              </TitleElement>
+            <TitleElement className={`${baseClass}--title`}>
+              {title}
+            </TitleElement>
+            {image && (
+              <div className={`${baseClass}--image--wrapper`}>
+                <picture>
+                  <img
+                    src={image}
+                    alt={title}
+                    className={`${baseClass}--picture`}
+                  />
+                </picture>
+              </div>
             )}
             {intro && <p className={`${baseClass}--intro`}>{intro}</p>}
-            {date && (
-              <time className={`${baseClass}--date`} dateTime={date.unix}>
-                {date.human}
-              </time>
+            {linklist && (
+              <LinkList
+                headline={linklist.headline}
+                linkgroup={linklist.linkgroup}
+              />
             )}
-            {details && <p className={`${baseClass}--date-extra`}>{details}</p>}
           </div>
         </div>
       </div>
@@ -139,4 +142,4 @@ const DetailCard = forwardRef<HTMLDivElement, DetailCardProps>(
   }
 );
 
-export { DetailCard };
+export { MultiLinkCard };
