--- packages/react/src/components/Cards/MultiLinkCard/index.tsx
@@ -1,1 +1,0 @@
-export * from "./MultiLinkCard";
