+++ packages/react/src/components/Cards/PromoCard.tsx
@@ -1,11 +1,17 @@
-import classNames from "classnames";
 import { forwardRef } from "react";
+import classNames from "classnames";
 
-import useGlobalSettings from "../../../hooks/useGlobalSettings";
-import { Profile, ProfileProps } from "../../Profile";
-import { CardSize, EventDate, HeadingTypes, ThemeTypes } from "../../../types";
+import useGlobalSettings from "../../hooks/useGlobalSettings";
+import {
+  CardCornerType,
+  CardSize,
+  HeadingTypes,
+  ThemeTypes,
+} from "../../types";
+import { LinkProps } from "../Link";
+import { Button } from "../Button";
 
-export type TextCardProps = {
+export type PromoCardProps = {
   /**
    * A line of text that appears as a small heading above the title of the card
    */
@@ -29,39 +35,45 @@ export type TextCardProps = {
   /**
    * How big should the card be
    */
-  size?: Omit<CardSize, "standard">;
+  size?: CardSize;
 
   /**
-   * Specify the event Date, in both human and Unix format.
+   * Apply an optional corner cut to the top of the card
    */
-  date?: EventDate;
+  cornercut?: CardCornerType;
 
   /**
-   * Profile information to be displayed on the card
+   * Introductory text in the card
    */
-  profile: ProfileProps;
+  intro?: string;
 
   /**
-   * Specify the URL for the card link
+   * Source link for the card
    */
   link?: string;
 
   /**
-   * Specify an optional className to be added to your TextCard.
+   * Call to action link
+   */
+  cta?: LinkProps;
+
+  /**
+   * Specify an optional className to be added to your PromoCard.
    */
   className?: string;
 };
 
-const TextCard = forwardRef<HTMLDivElement, TextCardProps>(
+const PromoCard = forwardRef<HTMLDivElement, PromoCardProps>(
   (
     {
       className,
       title,
       theme = "light",
       size = "narrow",
-      date,
-      profile,
+      cornercut,
+      intro,
       link,
+      cta,
       titleLevel: TitleElement = "p",
       eyebrow,
     },
@@ -71,11 +83,17 @@ const TextCard = forwardRef<HTMLDivElement, TextCardProps>(
 
     const baseClass = `${prefix}--card`;
 
-    const wrapperClass = classNames(`${baseClass}--wrapper`, className);
-    const cardClasses = classNames(baseClass, `${baseClass}__type__text`, {
+    const wrapperClass = classNames(
+      `${baseClass}--wrapper`,
+      className,
+      `${baseClass}--wrapper__type__promo`,
+      `${baseClass}--wrapper__type__promo__size__${size}`
+    );
+    const cardClasses = classNames(baseClass, `${baseClass}__type__promo`, {
       [`${baseClass}__action`]: link,
       [`${baseClass}__size__${String(size)}`]: size,
       [`${baseClass}__theme__${theme}`]: theme,
+      [`${baseClass}__cornercut`]: cornercut,
     });
 
     return (
@@ -90,20 +108,15 @@ const TextCard = forwardRef<HTMLDivElement, TextCardProps>(
               <TitleElement className={`${baseClass}--title`}>
                 {title}
               </TitleElement>
-              {date && (
-                <time className={`${baseClass}--date`} dateTime={date.unix}>
-                  {date.human}
-                </time>
-              )}
-              {profile && (
-                <Profile
-                  avatar={profile.avatar}
-                  description={profile.description}
-                  link={profile.link}
-                  name={profile.name}
-                  role={profile.role}
-                  theme={theme}
-                />
+              {intro && <p className={`${baseClass}--intro`}>{intro}</p>}
+              {cta?.label && (
+                <div className={`${baseClass}--cta`}>
+                  <Button
+                    link={{ url: cta.url, label: cta.label }}
+                    size="medium"
+                    type="primary"
+                  />
+                </div>
               )}
             </div>
           </div>
@@ -113,4 +126,4 @@ const TextCard = forwardRef<HTMLDivElement, TextCardProps>(
   }
 );
 
-export { TextCard };
+export { PromoCard };
