--- packages/react/src/components/Cards/PromoCard/index.tsx
@@ -1,1 +1,0 @@
-export * from "./PromoCard";
