--- packages/react/src/components/Cards/StatCard/index.tsx
@@ -1,1 +1,0 @@
-export * from "./StatCard";
