--- packages/react/src/components/Cards/TextCard/index.tsx
@@ -1,1 +1,0 @@
-export * from "./TextCard";
