+++ packages/react/src/components/index.ts
@@ -40,3 +40,4 @@ export { Navigation } from "./Navigation";
 export { CardGroup } from "./Cards/CardGroup";
 export { Breadcrumb } from "./Breadcrumb";
 export { Tabs } from "./Tabs";
+export * from "./Cards";
